🔀: Mocking 환경을 구축한다. (#47)
* 🛠 issue Template 수정

깃허브에서 수정한 커밋입니다.

* 🛠: pr-template 수정

깃허브에서 작성한 커밋입니다.

* 🛠: pr-template 수정

깃허브에서 수정한 커밋입니다.

* 🛠: pr-template 수정

깃허브에서 작성한 커밋입니다.

* 🧪 : jsdoc 설치 및 사용 방법 테스트

* ✨: jsdoc은 prop-types 라이브러리를 대신하는 라이브러리가 아니다.

* ✨: QuestionAndAnswer data를 json으로 가공 및 script에 적용

* 🧪 : axios로 인스턴한 get mocking페이지에서 적용 및 테스트

* 🐛: fetch.instance에서 발생한 BadRequest

* 🐛: 프로토컬 `https:`를 baseURL에 적용하지 않아서 발생한 Error

* ✨: fetchAPI를 활용한 custom hook 생성

* 🧪:  useFetchAPI를 테스트 해보고 log를 확인해보았다.

* ✨: API로 부터 받은 데이터들의 타입을 검출 할 수 있는 방법에 대한 주석처리

* ✨: dotevn로 환경 변수 세팅 완료 및 적용

* 🧪: fetchAPI.instance.test를 위해 생성한 test custom hook
