+++ package.json
@@ -18,7 +18,8 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "mocking": "json-server --watch src/Test/Mocking/db/vmwrResult.data.json --port 3001",
+    "mocking": "json-server --watch src/Test/Mocking/db/QuestionAndAnswer.data.json --port 3001",
+    "docs": "jsdoc -c jsdoc.config.json",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -54,6 +55,7 @@
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
+    "jsdoc": "^4.0.3",
     "json-server": "^1.0.0-beta.1",
     "prettier": "3.2.5",
     "sass": "^1.77.8"
