+++ src/Global/css/VmwrResult.component.scss
