+++ src/Test/CompileTest/TestApp.js
@@ -1,0 +1,25 @@
+import React from 'react'; // React를 import합니다.
+import logo from '../../src/Trash/logo.svg';
+
+const TestApp = () => {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.jsx</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
+  );
+};
+
+export default TestApp;
