+++ src/Test/CompileTest/myTest.js
@@ -1,0 +1,11 @@
+/**
+ * 이것은 테스트 입니다.
+ * A simple test function that logs 'test' to the console.
+ *
+ * @returns {void} This function does not return anything.
+ */
+const test = () => {
+  console.log('test');
+};
+
+export default test;
