+++ src/Test/Mocking/db/QuestionAndAnswer.data.json
@@ -1,0 +1,36 @@
+{
+  "results": [
+    {
+      "title": "미도인 성수 근로 결과지",
+      "detail": "상세보기",
+      "description": "상시 5인 미만 사업장에서 근무하시므로 추가적인 가산 수당이 없습니다. 주 근로 시간이 40시간이므로 주휴수당이 발생합니다. 5인 미만 사업장에서 근무하시기에 연장근로수당이 발생하지 않습니다. 5인 미만 사업장에서 근무하시기에 휴일근로수당이 발생하지 않습니다...",
+      "conditions": [
+        "주휴수당",
+        "기본요건",
+        "야간근로수당",
+        "연장근로수당"
+      ]
+    },
+    {
+      "title": "메가커피 근로 결과지",
+      "detail": "상세보기",
+      "description": "상시 5인 미만 사업장에서 근무하시므로 추가적인 가산 수당이 없습니다. 주 근로 시간이 40시간이므로 주휴수당이 발생합니다.",
+      "conditions": [
+        "주휴수당",
+        "기본요건",
+        "직장적합성",
+        "인간관계"
+      ]
+    },
+    {
+      "title": "아웃백 근로 결과지",
+      "detail": "상세보기",
+      "description": "상시 5인 미만 사업장에서 근무하시므로 추가적인 가산 수당이 없습니다.",
+      "conditions": [
+        "기본요건",
+        "추진적합성",
+        "주요관계"
+      ]
+    }
+  ]
+}
