--- src/Test/eslintTest/myTest.js
@@ -1,5 +1,0 @@
-const test = () => {
-  console.log('test');
-};
-
-export default test;
